smart-proctor-backend/requirements.txt
# --- Core Framework ---
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0

# --- Production Server ---
# Gunicorn manages Uvicorn workers in production
gunicorn>=21.2.0

# --- Security & Auth ---
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9

# --- Database ---
sqlalchemy>=2.0.29
# psycopg is the modern, faster driver for PostgreSQL (replaces psycopg2)
psycopg[binary]>=3.1.18
alembic>=1.13.1

# --- Utilities ---
requests>=2.31.0
python-dotenv>=1.0.1
email-validator>=2.1.1
# Tenacity is required for the "wait-for-db" script
tenacity>=8.2.3